+++ src/main/java/ResourceHandler.java
@@ -69,12 +69,12 @@ public class ResourceHandler {
     public static void addRandomEvents(JSONArray arr,GameState gameState) throws JSONException {
         JSONObject json;
         int duration;
-        double probability;
+        int probability;
         for(int i = 0; i < arr.length(); i++){
             json = arr.getJSONObject(i);
 
             duration = json.getInt("duration");
-            probability = json.getDouble("probability");
+            probability = json.getInt("probability");
             gameState.all_possible_events.add(new RandomEvent(duration, probability));
         }
     }
